% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bardata}
\alias{bardata}
\title{Simulated data for barcharts}
\format{
A data frame with 50 rows and 4 variables:
\describe{
  \item{STUDY}{Study number}
  \item{COUNT}{Count variable}
  \item{GROUP}{Group variable for stratification}
  \item{DOSE}{Dose amoung for stratification}
}
}
\usage{
bardata
}
\description{
Simulated data for barcharts
}
\keyword{datasets}
