% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{boxdata}
\alias{boxdata}
\title{Simulated data for box and whisker plots}
\format{
A data frame with 2000 rows and 6 variables:
\describe{
  \item{CONTINUOUS}{Continuous variable}
  \item{EXP}{Exponentiated values of CONTINUOUS}
  \item{CATEGORICAL}{Categorical variable}
  \item{GROUP}{Group variable for stratification}
  \item{SPLIT}{Another group variable for stratification}
  \item{DOSE}{Dose amoung for stratification}
}
}
\usage{
boxdata
}
\description{
Simulated data for box and whisker plots
}
\keyword{datasets}
