% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style.R
\name{cognigen_purple_style}
\alias{cognigen_purple_style}
\title{An alternative Cognigen ggplot2 style using a purple-hue ramp. It is
primarily intended to be used for order}
\usage{
cognigen_purple_style(n = 10, gray.first = FALSE)
}
\arguments{
\item{n}{number of values to be mapped}

\item{gray.first}{\code{logical} indicating whether the first group
of the discrete scale should be represented using a gray color}
}
\value{
\code{list} of style elements
}
\description{
An alternative Cognigen ggplot2 style using a purple-hue ramp. It is
primarily intended to be used for order
}
