% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-barcount.R, R/stat-barcount.R
\name{geom_barcount}
\alias{geom_barcount}
\alias{stat_barcount}
\title{Data count for barcounts}
\usage{
geom_barcount(
  mapping = NULL,
  data = NULL,
  stat = "barcount",
  position = "stack",
  ...,
  overall.stack = TRUE,
  digits = 3,
  width = NULL,
  na.rm = FALSE,
  orientation = NA,
  show.legend = NA,
  inherit.aes = TRUE
)

stat_barcount(
  mapping = NULL,
  data = NULL,
  geom = "barcount",
  position = "stack",
  ...,
  overall.stack = TRUE,
  digits = 3,
  width = NULL,
  na.rm = FALSE,
  orientation = NA,
  show.legend = NA,
  inherit.aes = TRUE
)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{\link[ggplot2:aes]{aes()}}. If specified and
\code{inherit.aes = TRUE} (the default), it is combined with the default mapping
at the top level of the plot. You must supply \code{mapping} if there is no plot
mapping.}

\item{data}{The data to be displayed in this layer. There are three
options:

If \code{NULL}, the default, the data is inherited from the plot
data as specified in the call to \code{\link[ggplot2:ggplot]{ggplot()}}.

A \code{data.frame}, or other object, will override the plot
data. All objects will be fortified to produce a data frame. See
\code{\link[ggplot2:fortify]{fortify()}} for which variables will be created.

A \code{function} will be called with a single argument,
the plot data. The return value must be a \code{data.frame}, and
will be used as the layer data. A \code{function} can be created
from a \code{formula} (e.g. \code{~ head(.x, 10)}).}

\item{position}{A position adjustment to use on the data for this layer. This
can be used in various ways, including to prevent overplotting and
improving the display. The \code{position} argument accepts the following:
\itemize{
\item The result of calling a position function, such as \code{position_jitter()}.
This method allows for passing extra arguments to the position.
\item A string naming the position adjustment. To give the position as a
string, strip the function name of the \code{position_} prefix. For example,
to use \code{position_jitter()}, give the position as \code{"jitter"}.
\item For more information and other ways to specify the position, see the
\link[ggplot2:layer_positions]{layer position} documentation.
}}

\item{...}{Other arguments passed on to \code{\link[ggplot2:layer]{layer()}}'s \code{params} argument. These
arguments broadly fall into one of 4 categories below. Notably, further
arguments to the \code{position} argument, or aesthetics that are required
can \emph{not} be passed through \code{...}. Unknown arguments that are not part
of the 4 categories below are ignored.
\itemize{
\item Static aesthetics that are not mapped to a scale, but are at a fixed
value and apply to the layer as a whole. For example, \code{colour = "red"}
or \code{linewidth = 3}. The geom's documentation has an \strong{Aesthetics}
section that lists the available options. The 'required' aesthetics
cannot be passed on to the \code{params}. Please note that while passing
unmapped aesthetics as vectors is technically possible, the order and
required length is not guaranteed to be parallel to the input data.
\item When constructing a layer using
a \verb{stat_*()} function, the \code{...} argument can be used to pass on
parameters to the \code{geom} part of the layer. An example of this is
\code{stat_density(geom = "area", outline.type = "both")}. The geom's
documentation lists which parameters it can accept.
\item Inversely, when constructing a layer using a
\verb{geom_*()} function, the \code{...} argument can be used to pass on parameters
to the \code{stat} part of the layer. An example of this is
\code{geom_area(stat = "density", adjust = 0.5)}. The stat's documentation
lists which parameters it can accept.
\item The \code{key_glyph} argument of \code{\link[ggplot2:layer]{layer()}} may also be passed on through
\code{...}. This can be one of the functions described as
\link[ggplot2:draw_key]{key glyphs}, to change the display of the layer in the legend.
}}

\item{overall.stack}{Defines whether an overall count is displayed for
stacked bars or if the counts of each individual component of the
stacked bars should be displayed.}

\item{digits}{Integer indicating the number of significant digits to be used.
Recognized values are \code{0..22}. Use \code{digits = 0} to display as
integers.}

\item{width}{Numeric value defining the width of the bars.}

\item{na.rm}{If \code{FALSE}, the default, missing values are removed with
a warning. If \code{TRUE}, missing values are silently removed.}

\item{orientation}{The orientation of the layer. The default (\code{NA})
automatically determines the orientation from the aesthetic mapping. In the
rare event that this fails it can be given explicitly by setting \code{orientation}
to either \code{"x"} or \code{"y"}. See the \emph{Orientation} section for more detail.}

\item{show.legend}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped.
\code{FALSE} never includes, and \code{TRUE} always includes.
It can also be a named logical vector to finely select the aesthetics to
display. To include legend keys for all levels, even
when no data exists, use \code{TRUE}.  If \code{NA}, all levels are shown in legend,
but unobserved levels are omitted.}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics,
rather than combining with them. This is most useful for helper functions
that define both data and aesthetics and shouldn't inherit behaviour from
the default plot specification, e.g. \code{\link[ggplot2:annotation_borders]{annotation_borders()}}.}

\item{geom, stat}{Use to override the default connection between
\code{geom_barcount} and \code{stat_barcount}.}
}
\description{
This function is intended to work in combination with
\code{\link[ggplot2]{geom_bar}} to display the sum of the values represented
by each bar. Like \code{\link[ggplot2]{geom_bar}}, this function works if the
\code{x} and/or \code{y} aesthetics are provided.
}
\section{Orientation}{
 
This geom treats each axis differently and, thus, can thus have two orientations. Often the orientation is easy to deduce from a combination of the given mappings and the types of positional scales in use. Thus, ggplot2 will by default try to guess which orientation the layer should have. Under rare circumstances, the orientation is ambiguous and guessing may fail. In that case the orientation can be specified directly using the \code{orientation} parameter, which can be either \code{"x"} or \code{"y"}. The value gives the axis that the geom should run along, \code{"x"} being the default orientation you would expect for the geom.
}

\section{Aesthetics}{

\code{geom_barcount()} understands the following aesthetics. Required aesthetics are displayed in bold and defaults are displayed for optional aesthetics:
\tabular{rll}{
 • \tab \strong{\code{\link[ggplot2:aes_position]{x}}} \tab   \cr
 • \tab \strong{\code{\link[ggplot2:aes_position]{y}}} \tab   \cr
 • \tab \strong{\code{label}} \tab   \cr
 • \tab \code{\link[ggplot2:aes_colour_fill_alpha]{alpha}} \tab → \code{NA} \cr
 • \tab \code{angle} \tab → \code{0} \cr
 • \tab \code{\link[ggplot2:aes_colour_fill_alpha]{colour}} \tab → \code{"black"} \cr
 • \tab \code{family} \tab → \code{""} \cr
 • \tab \code{fontface} \tab → \code{1} \cr
 • \tab \code{\link[ggplot2:aes_group_order]{group}} \tab → inferred \cr
 • \tab \code{hjust} \tab → \code{0.5} \cr
 • \tab \code{lineheight} \tab → \code{1.2} \cr
 • \tab \code{\link[ggplot2:aes_linetype_size_shape]{size}} \tab → \code{3} \cr
 • \tab \code{vjust} \tab → \code{0.5} \cr
}
Learn more about setting these aesthetics in \code{vignette("ggplot2-specs")}.
}

\examples{
library(ggplot2)

p <- ggplot(mpg)
p +
  aes(x = class) +
  geom_bar() +
  geom_barcount()

# Map class to y instead to flip the orientation
p +
  aes(y = class) +
  geom_bar() +
  geom_barcount()

# For stacked position
p <- ggplot(diamonds, aes(color, fill = cut))
p +
  geom_bar(position = 'stack') +
  geom_barcount()
p +
  geom_bar(position = 'stack') +
  geom_barcount(overall.stack = FALSE)

# For dodged position
p +
  geom_bar(position = 'dodge') +
  geom_barcount(position = position_dodge(width = 0.9))

# For fill position
p +
  geom_bar(position = 'fill') +
  geom_barcount(position = position_fill())

# For fillpercent position
p +
  geom_bar(position = 'fillpercent') +
  geom_barcount(position = position_fillpercent()) +
  ylab('count (\%)')
}
\seealso{
\code{\link[ggplot2]{geom_bar}}
}
