% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-boxplot2.R, R/stat-boxplot2.R
\name{geom_boxplot2}
\alias{geom_boxplot2}
\alias{stat_boxplot2}
\title{A variant on box and whiskers plot (in the style of Tukey, or not...)}
\usage{
geom_boxplot2(
  mapping = NULL,
  data = NULL,
  stat = "boxplot2",
  position = "dodge2",
  ...,
  median_symbol = TRUE,
  outlier.position = "jitter",
  outlier.colour = NULL,
  outlier.color = NULL,
  outlier.fill = NULL,
  outlier.shape = 21,
  outlier.size = 1.5,
  outlier.stroke = 0.5,
  outlier.alpha = NULL,
  whisker.cap = FALSE,
  notch = FALSE,
  notchwidth = 0.5,
  varwidth = FALSE,
  na.rm = FALSE,
  orientation = NA,
  show.legend = NA,
  inherit.aes = TRUE
)

stat_boxplot2(
  mapping = NULL,
  data = NULL,
  geom = "boxplot2",
  position = "dodge2",
  ...,
  coef = 1.5,
  na.rm = FALSE,
  orientation = NA,
  show.legend = NA,
  inherit.aes = TRUE
)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{\link[ggplot2:aes]{aes()}}. If specified and
\code{inherit.aes = TRUE} (the default), it is combined with the default mapping
at the top level of the plot. You must supply \code{mapping} if there is no plot
mapping.}

\item{data}{The data to be displayed in this layer. There are three
options:

If \code{NULL}, the default, the data is inherited from the plot
data as specified in the call to \code{\link[ggplot2:ggplot]{ggplot()}}.

A \code{data.frame}, or other object, will override the plot
data. All objects will be fortified to produce a data frame. See
\code{\link[ggplot2:fortify]{fortify()}} for which variables will be created.

A \code{function} will be called with a single argument,
the plot data. The return value must be a \code{data.frame}, and
will be used as the layer data. A \code{function} can be created
from a \code{formula} (e.g. \code{~ head(.x, 10)}).}

\item{position}{A position adjustment to use on the data for this layer. This
can be used in various ways, including to prevent overplotting and
improving the display. The \code{position} argument accepts the following:
\itemize{
\item The result of calling a position function, such as \code{position_jitter()}.
This method allows for passing extra arguments to the position.
\item A string naming the position adjustment. To give the position as a
string, strip the function name of the \code{position_} prefix. For example,
to use \code{position_jitter()}, give the position as \code{"jitter"}.
\item For more information and other ways to specify the position, see the
\link[ggplot2:layer_positions]{layer position} documentation.
}}

\item{...}{Other arguments passed on to \code{\link[ggplot2:layer]{layer()}}'s \code{params} argument. These
arguments broadly fall into one of 4 categories below. Notably, further
arguments to the \code{position} argument, or aesthetics that are required
can \emph{not} be passed through \code{...}. Unknown arguments that are not part
of the 4 categories below are ignored.
\itemize{
\item Static aesthetics that are not mapped to a scale, but are at a fixed
value and apply to the layer as a whole. For example, \code{colour = "red"}
or \code{linewidth = 3}. The geom's documentation has an \strong{Aesthetics}
section that lists the available options. The 'required' aesthetics
cannot be passed on to the \code{params}. Please note that while passing
unmapped aesthetics as vectors is technically possible, the order and
required length is not guaranteed to be parallel to the input data.
\item When constructing a layer using
a \verb{stat_*()} function, the \code{...} argument can be used to pass on
parameters to the \code{geom} part of the layer. An example of this is
\code{stat_density(geom = "area", outline.type = "both")}. The geom's
documentation lists which parameters it can accept.
\item Inversely, when constructing a layer using a
\verb{geom_*()} function, the \code{...} argument can be used to pass on parameters
to the \code{stat} part of the layer. An example of this is
\code{geom_area(stat = "density", adjust = 0.5)}. The stat's documentation
lists which parameters it can accept.
\item The \code{key_glyph} argument of \code{\link[ggplot2:layer]{layer()}} may also be passed on through
\code{...}. This can be one of the functions described as
\link[ggplot2:draw_key]{key glyphs}, to change the display of the layer in the legend.
}}

\item{median_symbol}{a logical value indicating whether to use a symbol
(\code{TRUE}) or a line (\code{FALSE}) to represent the median when a color
aesthetic is used.}

\item{outlier.position}{By default, outliers are displayed with a small degree of jitter. Sometimes
it can be useful to hide the outliers, for example when overlaying the raw
data points on top of the boxplot. Hiding the outliers can be achieved by
setting \code{outlier.position = NULL}. Importantly, this does not remove
the outliers, it only hides them, so the range calculated for the y-axis
will be the same with outliers shown and outliers hidden. If needed,
outliers can be displayed without jitter by setting \code{outlier.position
= 'identity'}.}

\item{outlier.colour, outlier.color, outlier.fill, outlier.shape, outlier.size, outlier.stroke, outlier.alpha}{Aesthetics for outliers inherited from the original
\code{\link[ggplot2]{geom_boxplot}} function but that are not used in
\code{geom_boxplot2}. Instead, outliers inherits colors, shapes, sizes from
the box aesthetics. These aesthetics were included to maintain code
compatibility with call to \code{\link[ggplot2]{geom_boxplot}}.}

\item{whisker.cap}{If \code{FALSE} (default), the whiskers are simple
segments. If \code{TRUE}, the end of the whiskers are delineated by
orthogonal segments.}

\item{notch}{If \code{FALSE} (default) make a standard box plot. If
\code{TRUE}, make a notched box plot. Notches are used to compare groups;
if the notches of two boxes do not overlap, this suggests that the medians
are significantly different.}

\item{notchwidth}{For a notched box plot, width of the notch relative to
the body (defaults to \code{notchwidth = 0.5}).}

\item{varwidth}{If \code{FALSE} (default) make a standard box plot. If
\code{TRUE}, boxes are drawn with widths proportional to the
square-roots of the number of observations in the groups (possibly
weighted, using the \code{weight} aesthetic).}

\item{na.rm}{If \code{FALSE}, the default, missing values are removed with
a warning. If \code{TRUE}, missing values are silently removed.}

\item{orientation}{The orientation of the layer. The default (\code{NA})
automatically determines the orientation from the aesthetic mapping. In the
rare event that this fails it can be given explicitly by setting \code{orientation}
to either \code{"x"} or \code{"y"}. See the \emph{Orientation} section for more detail.}

\item{show.legend}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped.
\code{FALSE} never includes, and \code{TRUE} always includes.
It can also be a named logical vector to finely select the aesthetics to
display. To include legend keys for all levels, even
when no data exists, use \code{TRUE}.  If \code{NA}, all levels are shown in legend,
but unobserved levels are omitted.}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics,
rather than combining with them. This is most useful for helper functions
that define both data and aesthetics and shouldn't inherit behaviour from
the default plot specification, e.g. \code{\link[ggplot2:annotation_borders]{annotation_borders()}}.}

\item{geom, stat}{Use to override the default connection between
\code{geom_boxplot2} and \code{stat_boxplot2}.}

\item{coef}{Length of the whiskers as multiple of IQR (if lower than 50) or a
confidence interval (if greater than or equal to 50). Defaults to 1.5.}
}
\description{
The boxplot compactly displays the distribution of a continuous variable. It
displays five summary statistics (the median, two hinges and two whiskers),
and all "outlying" points individually. \code{geom_boxplot2()} is a variant
on the \code{\link[ggplot2]{geom_boxplot}} function from the ggplot2 package.
It allows users to set whisker limits based upon a confidence interval rather
than a multiple of the IQR, allows to display outliers with jitter, and
provides slightly different graphical styles when grouping/coloring is used.
}
\section{Orientation}{
 
This geom treats each axis differently and, thus, can thus have two orientations. Often the orientation is easy to deduce from a combination of the given mappings and the types of positional scales in use. Thus, ggplot2 will by default try to guess which orientation the layer should have. Under rare circumstances, the orientation is ambiguous and guessing may fail. In that case the orientation can be specified directly using the \code{orientation} parameter, which can be either \code{"x"} or \code{"y"}. The value gives the axis that the geom should run along, \code{"x"} being the default orientation you would expect for the geom.
}

\section{Summary statistics}{

The lower and upper hinges correspond to the first and third quartiles (the
25th and 75th percentiles). This differs slightly from the method used by the
\code{\link[graphics]{boxplot}} function, and may be apparent with small
samples. See \code{\link[grDevices]{boxplot.stats}} for for more information
on how hinge positions are calculated for \code{\link[graphics]{boxplot}}.

By default, the upper whisker extends from the hinge to the largest value no
further than 1.5 * IQR from the hinge (where IQR is the inter-quartile range,
or distance between the first and third quartiles). The lower whisker extends
from the hinge to the smallest value at most 1.5 * IQR of the hinge. Data
beyond the end of the whiskers are called "outlying" points and are plotted
individually. If a \code{coef} argument is provided to the function call, the
whiskers may extend to alternative limits. If \code{coef} is set to a value
lower than 50, the value is used a multiplier to the IQR (default is 1.5 as
explained above). If \code{coef} is set to value greater than or equal to 50,
the whiskers extend to the limit of the \code{coef} confidence interval.

In a notched box plot, the notches extend \code{1.58 * IQR / sqrt(n)}. This
gives a roughly 95% confidence interval for comparing medians. See McGill et
al. (1978) for more details.
}

\section{Aesthetics}{

\code{geom_boxplot2()} understands the following aesthetics. Required aesthetics are displayed in bold and defaults are displayed for optional aesthetics:
\tabular{rll}{
 • \tab \strong{\code{\link[ggplot2:aes_position]{x}} \emph{or} \code{\link[ggplot2:aes_position]{y}}} \tab   \cr
 • \tab \strong{\code{lower} \emph{or} \code{xlower}} \tab   \cr
 • \tab \strong{\code{upper} \emph{or} \code{xupper}} \tab   \cr
 • \tab \strong{\code{middle} \emph{or} \code{xmiddle}} \tab   \cr
 • \tab \strong{\code{\link[ggplot2:aes_position]{ymin}} \emph{or} \code{\link[ggplot2:aes_position]{xmin}}} \tab   \cr
 • \tab \strong{\code{\link[ggplot2:aes_position]{ymax}} \emph{or} \code{\link[ggplot2:aes_position]{xmax}}} \tab   \cr
 • \tab \code{\link[ggplot2:aes_colour_fill_alpha]{alpha}} \tab → \code{NA} \cr
 • \tab \code{\link[ggplot2:aes_colour_fill_alpha]{colour}} \tab → \code{"grey20"} \cr
 • \tab \code{\link[ggplot2:aes_colour_fill_alpha]{fill}} \tab → \code{"transparent"} \cr
 • \tab \code{\link[ggplot2:aes_group_order]{group}} \tab → inferred \cr
 • \tab \code{\link[ggplot2:aes_linetype_size_shape]{linetype}} \tab → \code{"solid"} \cr
 • \tab \code{\link[ggplot2:aes_linetype_size_shape]{linewidth}} \tab → \code{0.5} \cr
 • \tab \code{\link[ggplot2:aes_linetype_size_shape]{shape}} \tab → \code{21} \cr
 • \tab \code{\link[ggplot2:aes_linetype_size_shape]{size}} \tab → \code{0.5} \cr
 • \tab \code{stroke} \tab → \code{0.5} \cr
 • \tab \code{weight} \tab → \code{1} \cr
}
Learn more about setting these aesthetics in \code{vignette("ggplot2-specs")}.
}

\section{Computed variables}{

\describe{
  \item{width}{width of boxplot}
  \item{ymin}{lower whisker = smallest observation greater than or equal to
  lower hinge - 1.5 * IQR or lower limit of the confidence interval}
  \item{lower}{lower hinge, 25\% quantile}
  \item{notchlower}{lower edge of notch = median - 1.58 * IQR / sqrt(n)}
  \item{middle}{median, 50\% quantile}
  \item{notchupper}{upper edge of notch = median + 1.58 * IQR / sqrt(n)}
  \item{upper}{upper hinge, 75\% quantile}
  \item{ymax}{upper whisker = largest observation less than or equal to upper
  hinge + 1.5 * IQR  or upper limit of the confidence interval}
}
}

\examples{
library(ggplot2)

p <- ggplot(mpg, aes(class, hwy))
p + geom_boxplot2()
p + geom_boxplot2(outlier.position = 'identity', coef = 90)
# Orientation follows the discrete axis
ggplot(mpg, aes(hwy, class)) + geom_boxplot2()

p + geom_boxplot2(notch = TRUE)
p + geom_boxplot2(whisker.cap = TRUE)
p + geom_boxplot2(varwidth = TRUE)
p + geom_boxplot2(fill = "white", colour = "#3366FF")

# Boxplots are automatically dodged when any aesthetic is a factor
p + geom_boxplot2(aes(colour = factor(drv)))

# Use the median_symbol argument to control how the median is drawn
p + geom_boxplot2(aes(colour = factor(drv)), median_symbol = FALSE)

# You can also use boxplots with continuous x, as long as you supply
# a grouping variable. cut_width is particularly useful
ggplot(diamonds, aes(carat, price)) +
  geom_boxplot2()
ggplot(diamonds, aes(carat, price)) +
  geom_boxplot2(aes(group = cut_width(carat, 0.25)))
# Adjust the transparency of outliers using outlier.alpha
ggplot(diamonds, aes(carat, price)) +
  geom_boxplot2(aes(group = cut_width(carat, 0.25)), outlier.alpha = 0.1)

# It's possible to draw a boxplot with your own computations if you
# use stat = "identity":
y <- rnorm(100)
df <- data.frame(
  x = 1,
  y0 = min(y),
  y25 = quantile(y, 0.25),
  y50 = median(y),
  y75 = quantile(y, 0.75),
  y100 = max(y)
)
ggplot(df, aes(x)) +
  geom_boxplot2(
   aes(ymin = y0, lower = y25, middle = y50, upper = y75, ymax = y100),
   stat = "identity"
 )
}
\references{
McGill, R., Tukey, J. W. and Larsen, W. A. (1978) Variations of
    box plots. The American Statistician 32, 12-16.
}
\seealso{
\code{\link[ggplot2]{geom_boxplot}} for original ggplot2 geom
  function.
}
