% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot-device-size.R
\name{get_device_size}
\alias{get_device_size}
\title{Generate width and height of a device}
\usage{
get_device_size(
  nplots = 1,
  layout = "portrait",
  units = c("in", "cm", "mm", "px"),
  dpi = 300
)
}
\arguments{
\item{nplots}{number of plots per page}

\item{layout}{either 'portrait' or 'landscape'}

\item{units}{one of ('in', 'cm', 'mm', 'px'). If not supplied, defaults to 'in'}

\item{dpi}{plot resolution}
}
\value{
a named \code{numeric} vector of width and height
}
\description{
Generate width and height of a device
}
\examples{
\dontrun{
dims <- get_device_size(nplots = 4,
                        units = 'in',
                        dpi = 300)

plot <- ggplot(mpg, aes(class, hwy)) +
  geom_point()

ggsave(filename = 'plot.png',
       plot = plot,
       width = dims['width'],
       height = dims['height'],
       units = 'in',
       dpi = 300)
}
}
