% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot-device-size.R
\name{get_device_size_pixel}
\alias{get_device_size_pixel}
\title{Generate width and height of a device in pixels}
\usage{
get_device_size_pixel(nplots = 1, layout = "portrait", res = 300)
}
\arguments{
\item{nplots}{number of plots per page}

\item{layout}{either 'portrait' or 'landscape'}

\item{res}{plot resolution}
}
\value{
a named \code{numeric} vector of width and height
}
\description{
Generate width and height of a device in pixels
}
