% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style.R
\name{get_style_colors}
\alias{get_style_colors}
\title{Get colors from ggplot2 style}
\usage{
get_style_colors(style = cognigen_style(), element = "scatter")
}
\arguments{
\item{style}{\code{list} of style elements, liked one return by \code{\link{cognigen_style}}
or \code{\link{read_style}}. Alternatively, \code{style} can be set to \code{'ggplot2'} to revert to the
defaults of the ggplot2 package.}

\item{element}{Either 'scatter' (default), 'ramp', 'bar', 'box.sym', 'box.rec', 'hist',
'hist.dens', 'vpc', 'vpc.style', 'vpc.tte.style', 'spline', 'hline', 'vline', 'abline', 'error',
'background'}
}
\value{
A vector of characters, typically RGB colors in hex format.
}
\description{
Extracts \code{color$col} from the user-defined level of a ggplot2 style.
}
\examples{
get_style_colors()
get_style_colors(element = 'bar')
get_style_colors(style = cognigen_purple_style())
}
\seealso{
\code{\link{cognigen_style}}
}
