% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-.R, R/geom-barcount.R, R/geom-boxcount.R,
%   R/geom-boxplot2.R, R/geom-crossbar2.R, R/geom-histcount.R,
%   R/position-fillpercent.R, R/stat-aaa.R, R/stat-barcount.R,
%   R/stat-boxcount.R, R/stat-boxplot2.R, R/stat-histcount.R
\docType{data}
\name{ggplot2-ggproto}
\alias{ggplot2-ggproto}
\alias{GeomBarcount}
\alias{GeomBoxcount}
\alias{GeomBoxplot2}
\alias{GeomCrossbar2}
\alias{GeomHistcount}
\alias{PositionFillpercent}
\alias{StatBin2}
\alias{StatCount2}
\alias{StatBarcount}
\alias{StatBoxcount}
\alias{StatBoxplot2}
\alias{StatHistcount}
\title{ggcognigen ggproto classes for ggplot2}
\description{
If you are creating a new geom, stat, position, or scale in another package,
you'll need to extend from `ggplot2::Geom`, `ggplot2::Stat`,
`ggplot2::Position`, or `ggplot2::Scale`. See \link[ggplot2]{ggproto}.
}
\keyword{datasets}
\keyword{internal}
