% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save.R
\name{ggsave_multiple}
\alias{ggsave_multiple}
\title{Save multiple ggplots}
\usage{
ggsave_multiple(
  filenames,
  plots = NULL,
  device = NULL,
  path = NULL,
  scale = 1,
  width = NA,
  height = NA,
  units = c("in", "cm", "mm"),
  dpi = 300,
  limitsize = TRUE,
  ...
)
}
\arguments{
\item{filenames}{A vector of file names to create on disk.}

\item{plots}{A list of plots to save.}

\item{device}{Device to use. Can either be a device function
(e.g. [png()]), or one of "eps", "ps", "tex" (pictex),
"pdf", "jpeg", "tiff", "png", "bmp", "svg" or "wmf" (windows only).}

\item{path}{Path of the directory to save plot to: `path` and `filename`
are combined to create the fully qualified file name. Defaults to the
working directory.}

\item{scale}{Multiplicative scaling factor.}

\item{width, height, units}{Plot size in `units` ("in", "cm", or "mm").
If not supplied, uses the size of current graphics device.}

\item{dpi}{Plot resolution. Also accepts a string input: "retina" (320),
"print" (300), or "screen" (72). Applies only to raster output types.}

\item{limitsize}{When `TRUE` (the default), `ggsave` will not
save images larger than 50x50 inches, to prevent the common error of
specifying dimensions in pixels.}

\item{...}{Other arguments passed on to the graphics device function,
as specified by `device`.}
}
\description{
`ggsave_multiple()` extends the capabilities of `ggplot2::ggsave()` by
allowing #' the saving of multiple ggplots to multiple image files. In
particular, it is suitable for saving multi-page plots created using
`ggforce::facet_grid_paginate()` or `ggforce::facet_wrap_paginate()`.
}
\details{
ggplots are to saved in separate files. Therefore, the `filenames` and
`plots` arguments must have the same length.

For multi-page ggplots, the filenames can be explicitly provided using
a C integer format expression, such as `figure%03d.png`, to produce
successive filenames `figure001.png`, `figure002.png`, etc. If the
that is not the case, `ggsave_multiple()` will automatically detect
which ggplot is a multi-page plot and amend the filename using the
format mentioned above. The number of digits used for page identification
will be depend on the number of pages to be created and will be 2
at the minimum.
}
\examples{
\dontrun{
library(ggplot2)
library(ggforce)

g1 <- ggplot(data = diamonds) +
  aes(x = carat, y = price) +
    geom_point()
g2 <- ggplot(data = diamonds) +
  aes(x = carat, y = price) +
  geom_point() +
  facet_wrap_paginate(vars(clarity), nrow = 2, ncol = 2, page = NULL)
g3 <- ggplot(diamonds) +
  aes(carat, price) +
  geom_point(alpha = 0.2) +
  facet_grid_paginate(color ~ cut:clarity, ncol = 3, nrow = 3, page = NULL)
ggsave_all(
  filenames = c('plot_g1.png', 'plot_g2.png', 'plot_g3.png'),
  plots = list(g1, g2, g3),
  path = '~/workspace_gitlab/tmp/'
}
}
