% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{histdata}
\alias{histdata}
\title{Simulated data for histograms}
\format{
A data frame with 1000 rows and 3 variables:
\describe{
  \item{RANDOM}{Random continuous variable}
  \item{GROUP}{Group variable for stratification}
  \item{DOSE}{Dose amoung for stratification}
}
}
\usage{
histdata
}
\description{
Simulated data for histograms
}
\keyword{datasets}
