% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kiwiexport.R
\name{kiwiexport_make}
\alias{kiwiexport_make}
\alias{make_docx}
\alias{make_html}
\alias{make_latex}
\alias{make_xlsx}
\title{Create export files (copied from kiwiexport)}
\arguments{
\item{...}{One or more huxtable or flextable (only for \code{make_docx})
objects.}

\item{file}{The path of the file to be created.}

\item{borders}{A single integer defining the width of the table borders.}

\item{orientation}{Defines the page orientation; either 'portrait' or
'landscape'.}

\item{options}{A list of LaTeX formatting options. The length of the list
must match the number of huxtable objects provided to \code{make_latex}.
Each list element should include the following levels:
\itemize{
\item header_rows A vector of integer defining the index of the header rows
in the huxtable object,
\item widths A numeric vector defining the relative widths of the table
columns,
\item orientation Defines the page orientation; either 'portrait' or 'landscape', and
\item pese A logical value defining whether the huxtable object is a
parameter estimates and standard errors table (TRUE) or not (FALSE).
}}
}
\description{
The \code{kiwiexport} packages provides utility functions to print huxtable
or flextable objects to either .html, .docx (Microsoft Office), .tex
(LaTeX), and .xlsx (Microsoft Excel) document.
}
\details{
All \code{make_...} functions do not return any value but all have the
side-effect of create a file (either .docx, .html, .tex, or .xlsx)
containing one or more tables.
}
\author{
Sebastien Bihorel
}
\keyword{methods}
