% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kiwiexport.R
\name{kiwiexport_theme_border}
\alias{kiwiexport_theme_border}
\alias{theme_border}
\alias{theme_border_flextable}
\title{Theme functions (copied from kiwiexport)}
\arguments{
\item{ht}{A huxtable object.}

\item{bold_rows}{The rows to be formatted with bold font face.}

\item{bold_cols}{The columns to be formatted with bold font face.}

\item{position}{The position of the table within the page. A length-one
character vector which may be "left", "center", "right", "wrapleft" or
"wrapright".}

\item{format}{Either "html", "word", or "latex"}

\item{header_rows}{A vector of integers defining the rows containing the
table headers.}
}
\value{
\code{theme_border} returns a formatted huxtable object.
\code{theme_border_flextable} returns a formatted flextable object, intended
for export into .docx (Microsoft Word) documents.
}
\description{
The \code{kiwiexport} packages provides theming functions to format tables.
The \code{theme_border} function is intended for huxtable objects, while the
\code{theme_border_flextable} transform a huxtable object into a flextable
object prior to applying the same theme.
}
\seealso{
\code{\link[huxtable]{position}},
}
\author{
Sebastien Bihorel
}
\keyword{methods}
