% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forest.R
\name{make_gmr_data}
\alias{make_gmr_data}
\title{Calculate and report geometric means (GM) and geometric mean ratios (GMR)}
\usage{
make_gmr_data(
  data,
  x_var,
  id_var,
  by = NULL,
  covariates,
  labels,
  ref_levels,
  paired = FALSE,
  var.equal = TRUE,
  ci = 0.9,
  sep = "\\t",
  digits = 3L,
  silent = FALSE
)
}
\arguments{
\item{data}{a data.frame of data provided in wide-form and expected to contain
one record per individual}

\item{x_var}{the variable(s) in \code{data} for which the GM and GMRs must be
calculated}

\item{id_var}{the variable in \code{data} differentiating individual data}

\item{by}{a stratification variable in \code{data}}

\item{covariates}{a vector of variables in \code{data}. They are expected to
be categorical variables (either logical, character, or factor) and define
the group for which the GM and GMRs of \code{x_var} are calculated}

\item{labels}{a vector of character or expression which defines the labels for
the variables set in \code{covariates}. Expected to have the same length as
\code{covariates}}

\item{ref_levels}{a vector of integer defining the reference level to be used
in the calculation of GMRs for the variables set in \code{covariates}. Expected
to have the same length as \code{covariates}}

\item{paired}{a logical value indicating if the results to be compared in the
calculation of GMRs are paired (TRUE) or not (FALSE, default)}

\item{var.equal}{a logical value indicating if equal variance on the linear
scale must be assumed (TRUE, default) or not (FALSE)}

\item{ci}{a numerical value between 0 and 1 defining the width of the
confidence interval around the calculated GMRs}

\item{sep}{the character that separates counts from GMR + confidence interval
in the \code{gmr_n_label} output variable}

\item{digits}{the number of significant digits to apply when creating result
labels (see Results)}

\item{silent}{a logical value indicating whether information messages should
be shown (FALSE) or hidden (TRUE)}
}
\value{
This function returns a data.frame with one row per variable in
\code{covariates} and level in these variables. The variables in this
data.frame are: \describe{
\item{x_var}{one of the variables defined in \code{x_var},}
\item{y_var}{one of the variables defined in \code{covariates},}
\item{y_label}{the label associated with each variables defined in
\code{covariates}; one of the label defined in \code{labels}, or, if this
argument was not defined, either the name of the variables and its 'label'
attributes in \code{data},}
\item{by}{one of the level of the \code{by} variable,}
\item{value}{one of the level of the \code{covariates} variables}
\item{n}{the number of individuals included in this particular level of
\code{y_var} and \code{by}; records with missing value of \code{x_var} are
excluded}
\item{reference}{a logical variable defining whether this particular
level of \code{y_var} is the reference for the calculation of GMRs}
\item{gm}{the GM value for this particular level of \code{y_var} and \code{by}}
\item{gm_lo}{the lower limit of the GM confidence interval}
\item{gm_hi}{the upper limit of the GM confidence interval}
\item{gm_label}{the GM label; structured as X1 [X2, X3], where X1
is \code{gm}, X2 is \code{gm_lo}, X3 is \code{gm_hi}, and
X1, X2, and X3 are formatted to \code{digits} significant digits}
\item{gmr}{the GMR value for this particular level of \code{y_var} and \code{by}
with respect to the reference level}
\item{gmr_lo}{the lower limit of the GMR confidence interval}
\item{gmr_hi}{the upper limit of the GMR confidence interval}
\item{gmr_label}{the GMR label; structured as X1 [X2, X3], where X1
is \code{gmr}, X2 is \code{gmr_lo}, X3 is \code{gmr_hi}, and
X1, X2, and X3 are formatted to \code{digits} significant digits}
\item{gmr_n_label}{the GMR label with n; structured as X1 [X2, X3] n = X4, where X1
is \code{gmr}, X2 is \code{gmr_lo}, X3 is \code{gmr_hi}, X4 is \code{n}, and
X1, X2, and X3 are formatted to \code{digits} significant digits}
}

\code{y_var}, \code{y_label}, and \code{value} are coerced to factors to
facilitate plotting using \code{make_forestplot}.
}
\description{
This function is intended to calculate and output geometric means (GM) and
geometric mean ratios (GMR) into a data.frame suitable for plotting using
the \code{make_forestplot} function.
}
\examples{
\dontrun{
# Use expo dataset provided in the ggcognigen package

make_gmr_data(
  data = expo,
  x_var = c('CMAXSS', 'AUCSS', 'CMINSS'),
  id_var = 'ID',
  by = NULL,
  covariates = c('AGE', 'WTKG', 'BMI', 'SEXF', 'RFCAT', 'CPCAT'),
  labels = c(
    expression('Age'~'(y)'),
    expression('Body'~'Weight'~'(kg)'),
    expression('Body'~'Mass'~'Index'~'(kg/'*m^2*')'),
    expression('Sex'),
    expression('Renal'~'Function'),
    expression('Hepatic'~'Function')
  ),
  ref_levels = c(2, 3, 2, 1, 1, 1),
  digits = 3,
  silent = TRUE
)

# same example with tidyverse syntax
expo \%>\%
  make_gmr_data(
    x_var = c(CMAXSS, AUCSS, CMINSS),
    id_var = ID,
    by = NULL,
    covariates = c(AGE, WTKG, BMI, SEXF, RFCAT, CPCAT),
    labels = c(
      expression('Age'~'(y)'),
      expression('Body'~'Weight'~'(kg)'),
      expression('Body'~'Mass'~'Index'~'(kg/'*m^2*')'),
      expression('Sex'),
      expression('Renal'~'Function'),
      expression('Hepatic'~'Function')
    ),
    ref_levels = c(2, 3, 2, 1, 1, 1),
    digits = 3,
    silent = TRUE
  )
}
}
