% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kiwiexport.R
\name{merge_repeated_cells}
\alias{merge_repeated_cells}
\title{Merge cells (copied from kiwiexport)}
\usage{
merge_repeated_cells(
  ht,
  row,
  col,
  no_merge_row = NULL,
  no_merge_col = NULL,
  merge_first = c("column", "row")
)
}
\arguments{
\item{ht}{A huxtable object.}

\item{row}{A vector of rows to be scanned for mergeable cells.}

\item{col}{A vector of columns to be scanned for mergeable cells.}

\item{no_merge_row}{A vector of rows within \code{row} which would not be
scanned for mergeable cells.}

\item{no_merge_col}{A vector of columns within \code{row} which would not be
scanned for mergeable cells.}

\item{merge_first}{Whether cells are merged by column (default) or row
first.  This affect how irregular shapes of identical cells are merged.}
}
\value{
\code{merge_repeated_cells} returns a huxtable object with merged
cells.
}
\description{
The \code{merge_repeated_cells} function automatically detects cells with
identical content in huxtables and merge theme. The cells must be contiguous
and form a matrix shape within the main table.
}
\author{
Sebastien Bihorel
}
\keyword{internal}
\keyword{methods}
