% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/position-fillpercent.R
\name{position_fillpercent}
\alias{position_fillpercent}
\title{Variant of ggplot2::position_fill}
\usage{
position_fillpercent(vjust = 1, reverse = FALSE)
}
\arguments{
\item{vjust}{Vertical adjustment for geoms that have a position
(like points or lines), not a dimension (like bars or areas). Set to
\code{0} to align with the bottom, \code{0.5} for the middle,
and \code{1} (the default) for the top.}

\item{reverse}{If \code{TRUE}, will reverse the default stacking order.
This is useful if you're rotating both the plot and legend.}
}
\description{
This function performs the same data transformation as \code{\link[ggplot2]{position_fill}},
but is normalized to 100 instead of 1.
}
\seealso{
\code{\link[ggplot2]{position_fill}}
}
