% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style.R
\name{read_style}
\alias{read_style}
\title{Read graphical style file created with the theme_maker Shiny app}
\usage{
read_style(path)
}
\arguments{
\item{path}{Relative or absolute path to the graphical style file}
}
\value{
\code{list} of style elements
}
\description{
Read graphical style file created with the theme_maker Shiny app
}
