% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot-style.R
\name{read_style_theme}
\alias{read_style_theme}
\title{Read graphical style file created with the theme_maker Shiny app}
\usage{
read_style_theme(file)
}
\arguments{
\item{file}{path to file}
}
\value{
\code{list} of style elements
}
\description{
Read graphical style file created with the theme_maker Shiny app
}
