% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot-style.R
\name{set_default_style}
\alias{set_default_style}
\title{Set default ggplot2 style when no aesthetics are used}
\usage{
set_default_style(style = cognigen_style())
}
\arguments{
\item{style}{\code{list} of style elements}
}
\value{
updates elements of default ggplot2 style then invisibly returns \code{NULL}
}
\description{
Set default ggplot2 style when no aesthetics are used
}
\examples{
\dontrun{
set_default_style()

ggplot(mpg, aes(class, hwy)) +
  geom_point()

}
}
