% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{theme_cognigen}
\alias{theme_cognigen}
\alias{theme_cognigen_grid}
\title{Cognigen ggplot2 themes. Run by default at startup of \code{ggcognigen}}
\usage{
theme_cognigen(smaller = FALSE, ...)

theme_cognigen_grid(
  major.x = TRUE,
  major.y = TRUE,
  minor.x = FALSE,
  minor.y = FALSE,
  smaller = FALSE,
  ...
)
}
\arguments{
\item{smaller}{\code{logical} indicating whether to use smaller font size}

\item{...}{additional theme elements to override defaults. See
\code{\link[ggplot2]{theme}}.}

\item{major.x, major.y, minor.x, minor.y}{\code{logical} indicating whether to
draw major/minor grid lines on the respective axis}
}
\description{
Cognigen ggplot2 themes. Run by default at startup of \code{ggcognigen}
}
\examples{
\dontrun{
p <- ggplot(mpg, aes(class, hwy)) +
  geom_point()

p + theme_cognigen()

p + theme_cognigen_grid()
}
}
