% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{xydata}
\alias{xydata}
\title{Simulated pharmacokinetic data for scatter and line plots}
\format{
A data frame with 720 rows and 5 variables:
\describe{
  \item{REP}{A simulation replicate index}
  \item{TIME}{Time since dose}
  \item{CONCENTRATION}{Drug concentration}
  \item{GROUP}{Dose group for stratification}
  \item{DOSE}{Dose amount for stratification}
}
}
\usage{
xydata
}
\description{
Simulated pharmacokinetic data for scatter and line plots
}
\keyword{datasets}
