% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style.R
\name{cognigen_style}
\alias{cognigen_style}
\title{Cognigen ggplot2 style (default graphical style)}
\usage{
cognigen_style()
}
\value{
\code{list} of style elements
}
\description{
Cognigen ggplot2 style (default graphical style)
}
