% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/axis.R
\name{format_continuous_cognigen}
\alias{format_continuous_cognigen}
\alias{major_breaks_log}
\alias{minor_breaks_log}
\title{Format continuous values}
\usage{
format_continuous_cognigen(x)

major_breaks_log(x)

minor_breaks_log(x)
}
\arguments{
\item{x}{breaks}
}
\value{
either a \code{numeric} or \code{expression} vector the same length as \code{x}
}
\description{
\code{format_continuous_cognigen} can be passed to the \code{labels}
argument of \code{scale_*_continuous} functions to format continuous values.

\code{major_breaks_log} and \code{minor_breaks_log} can be passed to
\code{major_breaks} and \code{minor_breaks} in \code{scale_*} functions to
format log10 axis scales, respectively.
}
\examples{
\dontrun{
min <- -5000
max <- 5000
set.seed(123)

data <- data.frame(
  x = runif(1000, min, max),
  y = runif(1000, min, max)
)

ggplot(data, aes(x = x, y = y)) +
  geom_point() +
  scale_y_continuous(labels = format_continuous_cognigen)

min <- 0
max.x <- 1000
max.y <- 3
set.seed(123)

data <- data.frame(
  x = runif(1000, min, max.x),
  y = rlnorm(1000, min, max.y)
)

ggplot(data, aes(x = x, y = y)) +
  geom_point() +
  theme_cognigen_grid(minor.x = TRUE, minor.y = TRUE) +
  scale_y_log10(minor_breaks = minor_breaks_log, labels = format_continuous_cognigen)
}
}
