% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-boxcount.R, R/stat-boxcount.R
\name{geom_boxcount}
\alias{geom_boxcount}
\alias{stat_boxcount}
\title{Data count for box and whiskers plot}
\usage{
geom_boxcount(
  mapping = NULL,
  data = NULL,
  stat = "boxcount",
  position = "dodge2",
  ...,
  outlier.position = "jitter",
  na.rm = FALSE,
  orientation = NA,
  show.legend = FALSE,
  inherit.aes = TRUE
)

stat_boxcount(
  mapping = NULL,
  data = NULL,
  geom = "boxcount",
  position = "dodge2",
  ...,
  coef = 1.5,
  na.rm = FALSE,
  orientation = NA,
  show.legend = FALSE,
  inherit.aes = TRUE
)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{\link[ggplot2:aes]{aes()}} or
\code{\link[ggplot2:aes_]{aes_()}}. If specified and \code{inherit.aes = TRUE} (the
default), it is combined with the default mapping at the top level of the
plot. You must supply \code{mapping} if there is no plot mapping.}

\item{data}{The data to be displayed in this layer. There are three
options:

If \code{NULL}, the default, the data is inherited from the plot
data as specified in the call to \code{\link[ggplot2:ggplot]{ggplot()}}.

A \code{data.frame}, or other object, will override the plot
data. All objects will be fortified to produce a data frame. See
\code{\link[ggplot2:fortify]{fortify()}} for which variables will be created.

A \code{function} will be called with a single argument,
the plot data. The return value must be a \code{data.frame}, and
will be used as the layer data. A \code{function} can be created
from a \code{formula} (e.g. \code{~ head(.x, 10)}).}

\item{position}{Position adjustment, either as a string, or the result of
a call to a position adjustment function.}

\item{...}{Other arguments passed on to \code{\link[ggplot2:layer]{layer()}}. These are
often aesthetics, used to set an aesthetic to a fixed value, like
\code{colour = "red"} or \code{size = 3}. They may also be parameters
to the paired geom/stat.}

\item{outlier.position}{This controls the placement of the counts above the boxes, depending on
whether outliers were hidden (\code{outlier.position = NULL}) or
displayed (\code{outlier.position = 'identity'} or \code{outlier.position = 'jitter'})
when the call to \code{\link{geom_boxplot2}} was made. By default, outliers are
assumed to be displayed.}

\item{na.rm}{If \code{FALSE}, the default, missing values are removed with
a warning. If \code{TRUE}, missing values are silently removed.}

\item{orientation}{The orientation of the layer. The default (\code{NA})
automatically determines the orientation from the aesthetic mapping. In the
rare event that this fails it can be given explicitly by setting \code{orientation}
to either \code{"x"} or \code{"y"}. See the \emph{Orientation} section for more detail.}

\item{show.legend}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped.
\code{FALSE} never includes, and \code{TRUE} always includes.
It can also be a named logical vector to finely select the aesthetics to
display.}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics,
rather than combining with them. This is most useful for helper functions
that define both data and aesthetics and shouldn't inherit behaviour from
the default plot specification, e.g. \code{\link[ggplot2:borders]{borders()}}.}

\item{geom, stat}{Use to override the default connection between
\code{geom_boxcount} and \code{stat_boxcount}.}

\item{coef}{Length of the whiskers as multiple of IQR (if lower than 50) or a confidence interval (if greater than or equal to 50). Defaults to 1.5.}
}
\description{
This function is intended to work in combination with \code{\link{geom_boxplot2}}
to display the number of data points used for the calculation of statistics
which are graphically represented by each box and whiskers.
}
\examples{
\dontrun{
p <- ggplot(mpg, aes(class, hwy))
p + geom_boxplot2() + geom_boxcount()

# For display on log axis scale, use the scale_y_continuous function
# Using coord_trans(y ='log10') would display the counts at the wrong place
p + geom_boxplot2() + geom_boxcount() + scale_y_continuous(trans = 'log10')
}
}
\seealso{
\code{\link{geom_boxplot2}}
}
