% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/device-size.R
\name{get_device_size}
\alias{get_device_size}
\alias{get_device_size_pixel}
\title{Generate width and height of a device based upon a given number of plots per
page}
\usage{
get_device_size(
  nplots = 1,
  layout = "portrait",
  units = c("in", "cm", "mm", "px"),
  dpi = 300
)

get_device_size_pixel(nplots = 1, layout = "portrait", dpi = 300)
}
\arguments{
\item{nplots}{Number of plots per page.}

\item{layout}{Either 'portrait' or 'landscape'.}

\item{units}{One of ('in', 'cm', 'mm', 'px'). If not supplied, defaults to
'in'.}

\item{dpi}{Plot resolution in dot per inch.}
}
\value{
A named \code{numeric} vector of width and height

A named \code{numeric} vector of width and height
}
\description{
Generate width and height of a device based upon a given number of plots per
page

Generate width and height in pixels of a device based upon a given number of plots per page
}
\examples{
\dontrun{
dims <- get_device_size(nplots = 4,
                        units = 'in',
                        dpi = 300)

plot <- ggplot(mpg, aes(class, hwy)) +
  geom_point()

ggsave(filename = 'plot.png',
       plot = plot,
       path = tempdir(),
       width = dims['width'],
       height = dims['height'],
       units = 'in',
       dpi = 300)
}
}
