% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-.R, R/cognigen-geom-boxplot.R,
%   R/cognigen-stat-boxplot.R
\docType{data}
\name{ggplot2-ggproto}
\alias{ggplot2-ggproto}
\alias{GeomCognigenBoxplot}
\alias{StatCognigenBoxplot}
\title{Cognigen ggproto classes for ggplot2}
\description{
If you are creating a new geom, stat, position, or scale in another package,
you'll need to extend from `ggplot2::Geom`, `ggplot2::Stat`,
`ggplot2::Position`, or `ggplot2::Scale`. See \link[ggplot2]{ggproto}.
}
\keyword{datasets}
\keyword{internal}
