% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save.R
\name{ggsave_multiple}
\alias{ggsave_multiple}
\title{Save multiple ggplots}
\usage{
ggsave_multiple(
  filenames,
  plots,
  device = NULL,
  path = NULL,
  scale = 1,
  width = NA,
  height = NA,
  units = c("in", "cm", "mm"),
  dpi = 300,
  limitsize = TRUE,
  ...
)
}
\arguments{
\item{filenames}{A vector of file names to create on disk.}

\item{plots}{A list of plots to save.}

\item{device}{Device to use. Can either be a device function
(e.g. \code{\link[grDevices]{png}}), or one of "eps", "ps", "tex" (pictex),
"pdf", "jpeg", "tiff", "png", "bmp", "svg" or "wmf" (windows only).}

\item{path}{Path of the directory to save plot to: \code{path} and
\code{filename} are combined to create the fully qualified file name.
Defaults to the working directory.}

\item{scale}{Multiplicative scaling factor.}

\item{width, height, units}{Plot size in \code{units} ("in", "cm", or "mm").
If not supplied, uses the size of current graphics device.}

\item{dpi}{Plot resolution. Also accepts a string input: "retina" (320),
"print" (300), or "screen" (72). Applies only to raster output types.}

\item{limitsize}{When \code{TRUE} (the default), \code{ggsave} will not
save images larger than 50x50 inches, to prevent the common error of
specifying dimensions in pixels.}

\item{...}{Other arguments passed on to the graphics device function,
as specified by \code{device}.}
}
\description{
\code{ggsave_multiple()} extends \code{\link[ggplot2]{ggsave}} to save
multiple ggplots to multiple files. In particular, it is suitable for saving
multi-page plots created using \code{\link[ggforce]{facet_grid_paginate}} or
\code{\link[ggforce]{facet_wrap_paginate}} from the ggforce package.

ggplots are saved in separate files. Therefore, the \code{filenames} and
\code{plots} arguments must have the same length.

For multi-page ggplots, the filenames can be explicitly provided using a C
integer format expression, such as \code{'figure\%03d.png'}, to produce
successive filenames \code{'figure001.png'}, \code{'figure002.png'}, etc. If
that is not the case, \code{ggsave_multiple()} will automatically detect
multi-page plots and amend filenames using the format mentioned above. The
number of digits used for page identification will depend on the number of
pages to be created and will be 2 at the minimum.
}
\examples{
library(ggplot2)
library(ggforce)

g1 <- ggplot(data = diamonds) +
  aes(x = carat, y = price) +
    geom_point()

g2 <- ggplot(data = diamonds) +
  aes(x = carat, y = price) +
  geom_point() +
  facet_wrap_paginate(vars(clarity), nrow = 2, ncol = 2, page = NULL)

g3 <- ggplot(diamonds) +
  aes(carat, price) +
  geom_point(alpha = 0.2) +
  facet_grid_paginate(color ~ cut, ncol = 3, nrow = 3, page = NULL)

g4 <- ggplot(data = diamonds) +
  aes(x = carat, y = price, color = clarity) +
  geom_point(alpha = 0.2)

g5 <- ggplot(data = diamonds) +
  aes(x = carat, y = price, color = cut) +
  geom_point(alpha = 0.2)

gs <- ggpubr::ggarrange(g1, g4, g5, nrow = 2, ncol = 2)

ggsave_multiple(
  filenames = c('plot_g1.png', 'plot_g2.png', 'plot_g3.png', 'plot_gs.png'),
  plots = list(g1, g2, g3, gs),
  path = tempdir()
)
}
