% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale.R
\name{scale_continuous_cognigen}
\alias{scale_continuous_cognigen}
\title{Continuous scale for color and fill aesthetics with a purple gradient}
\usage{
scale_continuous_cognigen(..., na.value = "grey50", guide = "colourbar")
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[ggplot2:continuous_scale]{ggplot2::continuous_scale}}
  \describe{
    \item{\code{scale_name}}{The name of the scale that should be used for error messages
associated with this scale.}
    \item{\code{name}}{The name of the scale. Used as the axis or legend title. If
\code{waiver()}, the default, the name of the scale is taken from the first
mapping used for that aesthetic. If \code{NULL}, the legend title will be
omitted.}
  }}

\item{na.value}{Colour to use for missing values}

\item{guide}{Type of legend. Use \code{"colourbar"} for continuous
colour bar, or \code{"legend"} for discrete colour legend.}
}
\description{
Continuous scale for color and fill aesthetics with a purple gradient
}
\examples{
\dontrun{
df <- data.frame(
  value = seq(1, 100),
  x = runif(100),
  y = runif(100),
  z = rnorm(100)
)

# color
ggplot(df, aes(x = x, y = y)) +
  geom_point(aes(color = z), pch = 19) +
  scale_continuous_cognigen()

# fill
ggplot(df, aes(x = value, y = y)) +
  geom_bar(aes(fill = z), stat = "identity") +
  scale_continuous_cognigen()
}
}
